# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Olena Studio",
    "version": "1.0.9",
    "description": "home project",
    "contact": {
      "name": "Maria Skydan",
      "email": "mariaskydan@gmail.com",
      "url": "https://github.com/maskaa22?page=1&tab=repositories"
    }
  },
...
  "paths": {
...
    "/auth/refresh": {
      "get": {
        "tags": [
          "auth"
        ],
        "operationId": "refresh",
        "summary": "Оновлення токена доступу",
        "description": "Після отримання cookies, перевіряється та оновлюються access та refresh token.",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/userSchema"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorSchemaIncorrectData"
            }
          },
          "401": {
            "description": "Error: Unauthorized",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUser"
            }
          }
        }
      }
    },
    "/auth": {
      "delete": {
        "tags": [
          "auth"
        ],
        "operationId": "deleteUser",
        "summary": "Видалення користувача",
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "type": "string",
            "description": "Введіть пошту користувача для видалення"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/deleteUserSchema"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFound"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUser"
            }
          }
        }
      }
    },
...
    "/auth/:user_id/reset-password": {
      "patch": {
        "tags": [
          "auth"
        ],
        "operationId": "resetPassword",
        "summary": "Новий пароль",
        "description": "На пошті в листі з відновленням пароля потрібно скопіювати частину між 'http://localhost:3000/:' та '/reset-password' та вставити в поле 'id'",
        "parameters": [
          {
            "name": "resetPassword",
            "in": "body",
            "description": "Вставте скопійований текст, введіть два рази новий пароль",
            "schema": {
              "type": "object",
              "default": {
                "_id": "642be61675e6c78ebc9c2dce",
                "password": "123456s",
                "passwordToo": "123456s"
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "type": "number",
              "example": 200
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "$ref": "#/definitions/errorValidatorResultSchema"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUser"
            }
          },
          "409": {
            "description": "Error: Conflict",
            "schema": {
              "$ref": "#/definitions/errorSchemaUnauthorized"
            }
          }
        }
      }
    },
    "/category": {
      "get": {
        "tags": [
          "category"
        ],
        "operationId": "getAllCategory",
        "summary": "Перегляд списку категорій товарів",
        "responses": {
          "200": {
            "description": "ОК",
            "schema": {
              "items": {
                "$ref": "#/definitions/categorySchema",
                "type": "array"
              }
            }
          },
          "408": {
            "description": "Error: Request Timeout",
            "schema": {
              "$ref": "#/definitions/errorSchemaTryAgainLater"
            }
          }
        }
      },
      "post": {
        "tags": [
          "category"
        ],
        "operationId": "createCategory",
        "summary": "Створення категорії",
        "parameters": [
          {
            "name": "category_name",
            "in": "body",
            "schema": {
              "type": "string",
              "default": {
                "category_name": "Маска"
              }
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "ОК",
            "schema": {
              "$ref": "#/definitions/categorySchema",
              "type": "object"
            }
          },
          "400": {
            "description": "Error: Bad Request",
            "schema": {
              "type": "object",
              "properties": {
                "massage": {
                  "type": "string",
                  "example": "Категорія вже існує"
                }
              }
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFound"
            }
          }
        }
      }
    },
...
    "/home/events/:user_id": {
      "get": {
        "tags": [
          "home"
        ],
        "operationId": "getCalendarEventForId",
        "summary": "Перегляд списку записів даного користувача по його id",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "type": "string",
            "required": true,
            "default": "6415b88bf74553185be6c3da",
            "description": "Ідентифікатор користувача"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/homeSchemaAllEvent"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUser"
            }
          }
        }
      }
    },
    "/home/find": {
      "get": {
        "tags": [
          "home"
        ],
        "operationId": "getFindEvent",
        "summary": "Список вільного часу для запису у заданий день( формат unix)",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "type": "string",
            "required": true,
            "default": "1678140000",
            "description": "Дата у форматі unix"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/homeSchemaForFind"
            }
          },
          "404": {
            "description": "Error: Not Found",
            "schema": {
              "$ref": "#/definitions/errorSchemaNotFoundUser"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "errorSchemaIncorrectData": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Некоректні дані"
        }
      }
    },
...
    "errorSchemaNotFound": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Заповніть поля"
        }
      }
    },
...
    "errorSchemaNotFoundUser": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Користувач відсутній"
        }
      }
    },
...
    "errorSchemaUnauthorized": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Не правильний логін чи пароль"
        }
      }
    },
    "errorSchemaTryAgainLater": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Будь ласка спробуйте пізніше"
        }
      }
    },
    "errorValidatorResultSchema": {
      "type": "object",
      "properties": {
        "errors": {
          "type": "object",
          "properties": {
            "value": {
              "type": "string",
              "example": "as2"
            },
            "msg": {
              "type": "string",
              "example": "Мінімальна довжина пароля 6 символів"
            },
            "param": {
              "type": "string",
              "example": "password"
            },
            "location": {
              "type": "string",
              "example": "body"
            }
          }
        },
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Некоректні дані"
        }
      }
    },
...
    "categorySchema": {
      "properties": {
        "_id": {
          "type": "string",
          "description": "Mongo ідентифікатор",
          "example": "555df2q5dcf"
        },
        "category_name": {
          "type": "string",
          "description": "Назва категорії",
          "example": "Маска"
        },
        "createdAt": {
          "description": "Дата створення категорії",
          "type": "string",
          "example": "2023-12-13T10:23:01.414Z"
        },
        "updatedAt": {
          "description": "Дата внесення змін",
          "type": "string",
          "example": "2023-12-13T10:24:38.892Z"
        },
        "__v": {
          "description": "Версійність в Mongo",
          "type": "integer",
          "example": 0
        },
        "id": {
          "type": "string",
          "example": "555df2q5dcf"
        }
      }
    },
...
    "homeSchemaAllEvent": {
      "type": "array",
      "items": {
        "properties": {
          "_id": {
            "type": "string",
            "description": "Mongo ідентифікатор",
            "example": "555df2q5dcf"
          },
          "title": {
            "type": "string",
            "description": "Ім'я особи яка записується на процедуру",
            "example": "Ванда"
          },
          "date": {
            "type": "string",
            "description": "Дата у форматі unix",
            "example": "1678140000"
          },
          "description": {
            "type": "string",
            "description": "Вибраний тип процедури",
            "example": "жіноча стрижка"
          },
          "time": {
            "type": "string",
            "description": "Час на який записана людина",
            "example": "14:00"
          },
          "user_id": {
            "type": "string",
            "description": "Ідентифікатор даного користувача",
            "example": "555dfgdfg546df2q5dcf"
          },
          "createdAt": {
            "description": "Дата створення запису",
            "type": "string",
            "example": "2023-12-13T10:23:01.414Z"
          },
          "updatedAt": {
            "description": "Дата внесення змін",
            "type": "string",
            "example": "2023-12-13T10:24:38.892Z"
          },
          "__v": {
            "description": "Версійність в Mongo",
            "type": "integer",
            "example": 0
          },
          "id": {
            "type": "string",
            "example": "555df2q5dcf"
          }
        }
      }
    },
    "homeSchemaForFind": {
      "type": "array",
      "items": {
        "properties": {
          "_id": {
            "type": "string",
            "description": "Mongo ідентифікатор",
            "example": "555df2q5dcf"
          },
          "title": {
            "type": "string",
            "description": "Вільний час для запису",
            "example": "08:00"
          },
          "id": {
            "type": "string",
            "example": "555df2q5dcf"
          }
        }
      }
    },
...
    "deleteUserSchema": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "description": "Текст повідомлення",
          "example": "Користувач видалений"
        }
      }
    },
...
    "userSchemaForUpdate": {
      "properties": {
        "_id": {
          "description": "Mongo ідентифікатор",
          "example": "555df2q5dcf"
        },
        "name": {
          "description": "Ім'я користувача",
          "type": "string",
          "example": "Луі"
        },
        "email": {
          "description": "Емеіл користувача",
          "type": "string",
          "example": "sds@example.com"
        },
        "role": {
          "description": "Роль користувача (user/admin)",
          "type": "string",
          "example": "user"
        },
        "foto": {
          "description": "Назва фотокартки",
          "type": "string",
          "example": "2dcsdvs-16sdfvdd-4b6d-925b-e3sgddfh6d7.jpg"
        },
        "is_active": {
          "description": "Підтвердження емейлу",
          "type": "boolean",
          "example": false
        },
        "createdAt": {
          "description": "Дата створення користувача",
          "type": "string",
          "example": "2023-12-13T10:23:01.414Z"
        },
        "updatedAt": {
          "description": "Дата внесення змін",
          "type": "string",
          "example": "2023-12-13T10:24:38.892Z"
        },
        "__v": {
          "description": "Версійність в Mongo",
          "type": "integer",
          "example": 0
        },
        "surname": {
          "description": "Прізвище користувача",
          "type": "string",
          "example": "Сальваторе"
        },
        "phone": {
          "description": "Номер телефону",
          "type": "string",
          "example": "0987793157"
        },
        "nameSity": {
          "description": "Місто одержання посилки для Нової пошти",
          "type": "string",
          "example": "Вінниця"
        },
        "nameDepartment": {
          "description": "Номер відділення Нової пошти",
          "type": "string",
          "example": "25"
        }
      }
    },
    "acessTokenSchema": {
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0MTU5MTA1Zjc0NsdgdhSGSEGjkk4MDQxODgzNCwiZXhwIjoxNjgwNDE5NzM0fQ.rCvXJn4CfpjWp6lOgB7rk-M02IdHKsXVEdrRr9KupJI"
    },
    "refreshTokenSchema": {
      "type": "string",
      "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpZCI6IjY0MTU5MTA1Zjc0NsdgdhSGSEGjkk4MDQxODgzNCwiZXhwIjoxNjgwNDE5NzM0fQ.rCvXJn4CfpjWp6lOgB7rk-M02IdHKsXVEdrRr9KupJI"
    },
    "userSchema": {
      "type": "array",
      "items": {
        "properties": {
          "user": {
            "type": "object",
            "$ref": "#/definitions/userSchemaForUpdate"
          },
          "access_token": {
            "type": "object",
            "$ref": "#/definitions/acessTokenSchema"
          },
          "refresh_token": {
            "type": "object",
            "$ref": "#/definitions/refreshTokenSchema"
          }
        }
      }
    },
...
  }
}